Don't create a separate class just for utility methods
