+++ checker/src/main/java/org/checkerframework/checker/index/samelen/SameLenVisitor.java
@@ -14,7 +14,6 @@ import org.checkerframework.common.basetype.BaseTypeChecker;
 import org.checkerframework.common.basetype.BaseTypeVisitor;
 import org.checkerframework.common.value.ValueCheckerUtils;
 import org.checkerframework.dataflow.expression.JavaExpression;
-import org.checkerframework.dataflow.expression.JavaExpressions;
 import org.checkerframework.framework.type.AnnotatedTypeMirror;
 import org.checkerframework.javacutil.TreeUtils;
 
@@ -41,7 +40,7 @@ public class SameLenVisitor extends BaseTypeVisitor<SameLenAnnotatedTypeFactory>
                 && !(valueType.hasAnnotation(PolySameLen.class)
                         && varType.hasAnnotation(PolySameLen.class))) {
 
-            JavaExpression rhs = JavaExpressions.fromTree(atypeFactory, (ExpressionTree) valueTree);
+            JavaExpression rhs = JavaExpression.fromTree(atypeFactory, (ExpressionTree) valueTree);
             if (rhs != null && SameLenAnnotatedTypeFactory.mayAppearInSameLen(rhs)) {
                 String rhsExpr = rhs.toString();
                 AnnotationMirror am = valueType.getAnnotation(SameLen.class);
