+++ checker/src/main/java/org/checkerframework/checker/index/searchindex/SearchIndexTransfer.java
@@ -8,7 +8,7 @@ import org.checkerframework.checker.index.qual.SearchIndexFor;
 import org.checkerframework.common.value.ValueCheckerUtils;
 import org.checkerframework.dataflow.analysis.TransferInput;
 import org.checkerframework.dataflow.cfg.node.Node;
-import org.checkerframework.dataflow.expression.JavaExpressions;
+import org.checkerframework.dataflow.expression.JavaExpression;
 import org.checkerframework.framework.flow.CFAnalysis;
 import org.checkerframework.framework.flow.CFStore;
 import org.checkerframework.framework.flow.CFValue;
@@ -65,7 +65,7 @@ public class SearchIndexTransfer extends IndexAbstractTransfer {
                 List<String> arrays =
                         ValueCheckerUtils.getValueOfAnnotationWithStringArgument(rightSI);
                 AnnotationMirror nif = aTypeFactory.createNegativeIndexFor(arrays);
-                store.insertValue(JavaExpressions.fromNode(analysis.getTypeFactory(), right), nif);
+                store.insertValue(JavaExpression.fromNode(analysis.getTypeFactory(), right), nif);
             }
         }
     }
