+++ checker/src/main/java/org/checkerframework/checker/nullness/NullnessTransfer.java
@@ -31,7 +31,6 @@ import org.checkerframework.dataflow.cfg.node.NullLiteralNode;
 import org.checkerframework.dataflow.cfg.node.ReturnNode;
 import org.checkerframework.dataflow.cfg.node.ThrowNode;
 import org.checkerframework.dataflow.expression.JavaExpression;
-import org.checkerframework.dataflow.expression.JavaExpressions;
 import org.checkerframework.dataflow.expression.LocalVariable;
 import org.checkerframework.framework.flow.CFAbstractAnalysis;
 import org.checkerframework.framework.flow.CFAbstractStore;
@@ -114,7 +113,7 @@ public class NullnessTransfer
      * implement case 2.
      */
     protected void makeNonNull(NullnessStore store, Node node) {
-        JavaExpression internalRepr = JavaExpressions.fromNode(nullnessTypeFactory, node);
+        JavaExpression internalRepr = JavaExpression.fromNode(nullnessTypeFactory, node);
         store.insertValue(internalRepr, NONNULL);
     }
 
@@ -184,7 +183,7 @@ public class NullnessTransfer
             List<Node> secondParts = splitAssignments(secondNode);
             for (Node secondPart : secondParts) {
                 JavaExpression secondInternal =
-                        JavaExpressions.fromNode(nullnessTypeFactory, secondPart);
+                        JavaExpression.fromNode(nullnessTypeFactory, secondPart);
                 if (CFAbstractStore.canInsertJavaExpression(secondInternal)) {
                     thenStore = thenStore == null ? res.getThenStore() : thenStore;
                     elseStore = elseStore == null ? res.getElseStore() : elseStore;
@@ -369,7 +368,7 @@ public class NullnessTransfer
         // Refine result to @NonNull if n is an invocation of Map.get, the argument is a key for
         // the map, and the map's value type is not @Nullable.
         if (keyForTypeFactory != null && keyForTypeFactory.isMapGet(n)) {
-            String mapName = JavaExpressions.fromNode(nullnessTypeFactory, receiver).toString();
+            String mapName = JavaExpression.fromNode(nullnessTypeFactory, receiver).toString();
             AnnotatedTypeMirror receiverType = nullnessTypeFactory.getReceiverType(n.getTree());
 
             if (keyForTypeFactory.isKeyForMap(mapName, methodArgs.get(0))
