+++ checker/src/main/java/org/checkerframework/checker/optional/OptionalVisitor.java
@@ -24,7 +24,6 @@ import org.checkerframework.common.basetype.BaseTypeChecker;
 import org.checkerframework.common.basetype.BaseTypeValidator;
 import org.checkerframework.common.basetype.BaseTypeVisitor;
 import org.checkerframework.dataflow.expression.JavaExpression;
-import org.checkerframework.dataflow.expression.JavaExpressions;
 import org.checkerframework.framework.type.AnnotatedTypeFactory;
 import org.checkerframework.framework.type.AnnotatedTypeMirror.AnnotatedDeclaredType;
 import org.checkerframework.javacutil.TreeUtils;
@@ -160,8 +159,8 @@ public class OptionalVisitor
 
     /** Return true if the two trees represent the same expression. */
     private boolean sameExpression(ExpressionTree tree1, ExpressionTree tree2) {
-        JavaExpression r1 = JavaExpressions.fromTree(atypeFactory, tree1);
-        JavaExpression r2 = JavaExpressions.fromTree(atypeFactory, tree1);
+        JavaExpression r1 = JavaExpression.fromTree(atypeFactory, tree1);
+        JavaExpression r2 = JavaExpression.fromTree(atypeFactory, tree1);
         if (r1 != null && !r1.containsUnknown() && r2 != null && !r2.containsUnknown()) {
             return r1.equals(r2);
         } else {
