+++ framework/src/main/java/org/checkerframework/common/aliasing/AliasingTransfer.java
@@ -16,7 +16,6 @@ import org.checkerframework.dataflow.cfg.node.MethodInvocationNode;
 import org.checkerframework.dataflow.cfg.node.Node;
 import org.checkerframework.dataflow.cfg.node.ObjectCreationNode;
 import org.checkerframework.dataflow.expression.JavaExpression;
-import org.checkerframework.dataflow.expression.JavaExpressions;
 import org.checkerframework.framework.flow.CFAbstractAnalysis;
 import org.checkerframework.framework.flow.CFStore;
 import org.checkerframework.framework.flow.CFTransfer;
@@ -67,7 +66,7 @@ public class AliasingTransfer extends CFTransfer {
             return super.visitAssignment(n, in); // Do normal refinement.
         }
         // Widen the type of the rhs if the RHS's declared type wasn't @Unique.
-        JavaExpression rhsExpr = JavaExpressions.fromNode(factory, rhs);
+        JavaExpression rhsExpr = JavaExpression.fromNode(factory, rhs);
         in.getRegularStore().clearValue(rhsExpr);
         return new RegularTransferResult<>(null, in.getRegularStore());
     }
@@ -105,7 +104,7 @@ public class AliasingTransfer extends CFTransfer {
             AnnotatedTypeMirror paramType = paramTypes.get(i);
             if (!paramType.hasAnnotation(NonLeaked.class)
                     && !paramType.hasAnnotation(LeakedToResult.class)) {
-                store.clearValue(JavaExpressions.fromNode(factory, arg));
+                store.clearValue(JavaExpression.fromNode(factory, arg));
             }
         }
 
@@ -115,7 +114,7 @@ public class AliasingTransfer extends CFTransfer {
         if (receiverType != null
                 && !receiverType.hasAnnotation(LeakedToResult.class)
                 && !receiverType.hasAnnotation(NonLeaked.class)) {
-            store.clearValue(JavaExpressions.fromNode(factory, receiver));
+            store.clearValue(JavaExpression.fromNode(factory, receiver));
         }
     }
 
@@ -150,7 +149,7 @@ public class AliasingTransfer extends CFTransfer {
                 if (factory.getAnnotatedType(param).hasAnnotation(LeakedToResult.class)) {
                     // If argument can leak to result, and parent is not a
                     // single statement, remove that node from store.
-                    store.clearValue(JavaExpressions.fromNode(factory, arg));
+                    store.clearValue(JavaExpression.fromNode(factory, arg));
                 }
             }
 
@@ -159,7 +158,7 @@ public class AliasingTransfer extends CFTransfer {
             AnnotatedExecutableType annotatedType = factory.getAnnotatedType(methodElement);
             AnnotatedDeclaredType receiverType = annotatedType.getReceiverType();
             if (receiverType != null && receiverType.hasAnnotation(LeakedToResult.class)) {
-                store.clearValue(JavaExpressions.fromNode(factory, receiver));
+                store.clearValue(JavaExpression.fromNode(factory, receiver));
             }
         }
         // If parent is a statement, processPostconditions will handle the
