+++ framework/src/main/java/org/checkerframework/common/value/ValueTransfer.java
@@ -55,7 +55,6 @@ import org.checkerframework.dataflow.cfg.node.StringConversionNode;
 import org.checkerframework.dataflow.cfg.node.StringLiteralNode;
 import org.checkerframework.dataflow.cfg.node.UnsignedRightShiftNode;
 import org.checkerframework.dataflow.expression.JavaExpression;
-import org.checkerframework.dataflow.expression.JavaExpressions;
 import org.checkerframework.dataflow.expression.Unknown;
 import org.checkerframework.dataflow.util.NodeUtils;
 import org.checkerframework.framework.flow.CFAbstractAnalysis;
@@ -536,7 +535,7 @@ public class ValueTransfer extends CFTransfer {
      * or @ArrayLenRange annotation for the array or string.
      */
     private void refineAtLengthAccess(Node lengthNode, Node receiverNode, CFStore store) {
-        JavaExpression lengthExpr = JavaExpressions.fromNode(analysis.getTypeFactory(), lengthNode);
+        JavaExpression lengthExpr = JavaExpression.fromNode(analysis.getTypeFactory(), lengthNode);
 
         // If the expression is not representable (for example if String.length() for some reason is
         // not marked @Pure, then do not refine.
@@ -556,7 +555,7 @@ public class ValueTransfer extends CFTransfer {
         if (AnnotationUtils.areSameByName(lengthAnno, ValueAnnotatedTypeFactory.BOTTOMVAL_NAME)) {
             // If the length is bottom, then this is dead code, so the receiver type
             // should also be bottom.
-            JavaExpression receiver = JavaExpressions.fromNode(atypeFactory, receiverNode);
+            JavaExpression receiver = JavaExpression.fromNode(atypeFactory, receiverNode);
             store.insertValue(receiver, lengthAnno);
             return;
         }
@@ -583,7 +582,7 @@ public class ValueTransfer extends CFTransfer {
         } else {
             combinedRecAnno = hierarchy.greatestLowerBound(oldRecAnno, newRecAnno);
         }
-        JavaExpression receiver = JavaExpressions.fromNode(analysis.getTypeFactory(), receiverNode);
+        JavaExpression receiver = JavaExpression.fromNode(analysis.getTypeFactory(), receiverNode);
         store.insertValue(receiver, combinedRecAnno);
     }
 
@@ -1380,7 +1379,7 @@ public class ValueTransfer extends CFTransfer {
     private void addAnnotationToStore(CFStore store, AnnotationMirror anno, Node node) {
         // If node is assignment, iterate over lhs and rhs; otherwise, iterator contains just node.
         for (Node internal : splitAssignments(node)) {
-            JavaExpression je = JavaExpressions.fromNode(analysis.getTypeFactory(), internal);
+            JavaExpression je = JavaExpression.fromNode(analysis.getTypeFactory(), internal);
             CFValue currentValueFromStore;
             if (CFAbstractStore.canInsertJavaExpression(je)) {
                 currentValueFromStore = store.getValue(je);
@@ -1537,7 +1536,7 @@ public class ValueTransfer extends CFTransfer {
             // Update the annotation of the receiver
             if (minLength != 0) {
                 JavaExpression receiver =
-                        JavaExpressions.fromNode(atypeFactory, n.getTarget().getReceiver());
+                        JavaExpression.fromNode(atypeFactory, n.getTarget().getReceiver());
 
                 AnnotationMirror minLenAnno =
                         atypeFactory.createArrayLenRangeAnnotation(minLength, Integer.MAX_VALUE);
